requirements.txt
pyTelegramBotAPI
wikipedia
Pillow
requests
python-dotenv
peewee